# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Free Roulette Spin System API
  version: 1.0.0
  description: |
    ## Overview
    
    The Free Roulette Spin System API provides endpoints for managing user rewards, 
    roulette gameplay, and letter collection mechanics in a casino platform.
    
    ## Features
    
    - **Mission Management**: Earn and claim free spins through deposit-based missions
    - **Roulette Gameplay**: Spin the roulette wheel to win cash and letters
    - **Letter Collection**: Collect letters to form words and claim collection bonuses
    - **Deposit Processing**: Process deposits and evaluate mission eligibility
    - **Transaction Tracking**: Complete audit trail of all balance changes
    
    ## Authentication
    
    All endpoints require a valid User ID in the `X-User-Id` header. This ID should 
    correspond to a user in the main casino system.
    
    ## Error Handling
    
    The API uses standard HTTP status codes and returns structured error responses 
    with detailed error information including error codes and user-friendly messages.
  contact:
    name: Casino Development Team
    email: dev@casino.com
    url: https://casino.com
  license:
    name: Proprietary
    url: https://casino.com/license
servers:
  - url: http://localhost:8080/api
    description: Development server - Local development environment
  - url: https://staging-api.casino.com/roulette
    description: Staging server - Testing environment
...
security:
  - UserIdHeader: []
paths:
...
  /deposits:
    post:
      tags:
        - Deposits
      summary: Process user deposit
      description: Process a user deposit, update their balance, and evaluate eligible missions for free spins.
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
            examples:
              deposit_request:
                summary: Deposit request
                value:
                  amount: 100.00
      responses:
        '200':
          description: Deposit processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositResponse'
              examples:
                successful_deposit:
                  summary: Successful deposit
                  value:
                    success: true
                    message: "Deposit processed successfully"
                    depositAmount: 100.00
                    newBalance: 250.00
                    eligibleMissions: 2
                    missionIds: [1, 2]
        '400':
          description: Invalid deposit amount
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_amount:
                  summary: Invalid amount
                  value:
                    success: false
                    error: "INVALID_DEPOSIT_AMOUNT"
                    message: "Deposit amount must be greater than 0"
                    timestamp: "2024-01-15T10:30:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /deposits/history:
    get:
      tags:
        - Deposits
      summary: Get deposit history
      description: Retrieve the user's deposit transaction history.
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      responses:
        '200':
          description: Successfully retrieved deposit history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositHistoryResponse'
              examples:
                deposit_history:
                  summary: Deposit history
                  value:
                    deposits:
                      - id: 1
                        userId: 12345
                        transactionType: "DEPOSIT"
                        amount: 100.00
                        description: "User deposit"
                        createdAt: "2024-01-15T10:30:00Z"
                    totalDeposited: 100.00
                    depositCount: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    UserIdHeader:
      type: apiKey
      in: header
      name: X-User-Id
      description: User ID header for authentication. Must be a valid positive integer.
  parameters:
    UserIdHeader:
      name: X-User-Id
      in: header
      required: true
      description: User ID
      schema:
        type: integer
        format: int64
        minimum: 1
        example: 12345
  schemas:
...
    DepositRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: decimal
          minimum: 0.01
          description: Deposit amount
          example: 100.00
    DepositResponse:
      type: object
      required:
        - success
        - message
        - depositAmount
        - newBalance
        - eligibleMissions
        - missionIds
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Deposit processed successfully"
        depositAmount:
          type: number
          format: decimal
          description: Amount deposited
          example: 100.00
        newBalance:
          type: number
          format: decimal
          description: User's new balance after deposit
          example: 250.00
        eligibleMissions:
          type: integer
          minimum: 0
          description: Number of missions eligible for this deposit
          example: 2
        missionIds:
          type: array
          items:
            type: integer
            format: int64
          description: IDs of eligible missions
          example: [1, 2]
    DepositHistoryResponse:
      type: object
      required:
        - deposits
        - totalDeposited
        - depositCount
      properties:
        deposits:
          type: array
          items:
            $ref: '#/components/schemas/TransactionLog'
          description: List of deposit transactions
        totalDeposited:
          type: number
          format: decimal
          description: Total amount deposited
          example: 100.00
        depositCount:
          type: integer
          minimum: 0
          description: Number of deposits made
          example: 1
    TransactionLog:
      type: object
      required:
        - id
        - userId
        - transactionType
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: Transaction ID
          example: 1
        userId:
          type: integer
          format: int64
          description: User ID
          example: 12345
        transactionType:
          type: string
          description: Type of transaction
          example: "DEPOSIT"
        amount:
          type: number
          format: decimal
          description: Transaction amount
          example: 100.00
        description:
          type: string
          description: Transaction description
          example: "User deposit"
        createdAt:
          type: string
          format: date-time
          description: Transaction timestamp
          example: "2024-01-15T10:30:00Z"
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error code
          example: "INSUFFICIENT_SPINS"
        message:
          type: string
          description: Human-readable error message
          example: "You don't have enough spins available"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
  responses:
    BadRequest:
      description: Bad Request - Invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_user_id:
              summary: Invalid User ID
              value:
                success: false
                error: "INVALID_USER_ID"
                message: "User ID must be a positive integer"
                timestamp: "2024-01-15T10:30:00Z"
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_header:
              summary: Missing X-User-Id header
              value:
                success: false
                error: "MISSING_USER_ID"
                message: "X-User-Id header is required"
                timestamp: "2024-01-15T10:30:00Z"
    InternalServerError:
      description: Internal Server Error - Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            server_error:
              summary: Internal server error
              value:
                success: false
                error: "INTERNAL_ERROR"
                message: "An unexpected error occurred. Please try again later."
                timestamp: "2024-01-15T10:30:00Z"
...